Allow the version to be check using a custom JSON file online.
Add a `customVersionFileURLPath` property to check a custom versions file instead of the App Store version. If this property is set, Siren will not check the AppStore for the latest version, but will first check this file. The online JSON file can contain two version strings: "notice" and "minimal".
If the current version is older than the _minimal_ version, a `.Force` type alert is shown. If the current version is the same or older than the _notice_ version, a `.Option` type alert is shown.