Add delegate method `sirenDidCheckNoUpdateNecessary`.
This delegate method is called when the check is performed without any (real) error, and it concluded that there is no new version available. We need this delegate method to be able to construct a mechanism that keeps the user locked with a Force alert, when they have a version, that is older than the required minimum.